many changes
I've been working on patching many bugs as well as adding new functionality.

The codebase is large enough that I feel I should start version control. I'll be using "Semantic Versioning", with an extra stipulation that versions under 0.25 are alpha testing and versions greater than and including 0.5 but less than 1.0 are beta (for the compiler as a whole).

I'm also going to add unit tests and try to achieve close to full code coverage.